refractor: safe access, update guards
