+++ Cargo.toml
@@ -1,4 +1,4 @@
 [workspace]
 members = [
-    "src",
+    "wicp",
 ]
