+++ dfx.json
@@ -1,21 +1,16 @@
 {
-    "canisters": {
-      "wicp": {
-        "type": "custom",
-        "candid": "src/wicp.did",
-        "wasm": "target/wasm32-unknown-unknown/release/opt.wasm"
-      },
-      "wicp_test": {
-         "type": "custom",
-         "candid": "src/wicp.did",
-         "wasm": "target/wasm32-unknown-unknown/release/opt.wasm"
-      }
-    },
-    "networks": {
-      "local": {
-        "bind": "127.0.0.1:8000",
-        "type": "ephemeral"
-      }
-    },
-    "version": 1
+	"canisters": {
+		"wicp": {
+			"package": "wicp",
+			"candid": "wicp/wicp.did",
+			"type": "rust"
+		}
+	},
+	"networks": {
+		"local": {
+			"bind": "127.0.0.1:8000",
+			"type": "ephemeral"
+		}
+	},
+	"version": 1
 }
