+++ wicp/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "rust"
-version = "0.1.0"
+name = "wicp"
+version = "2.0.0"
 edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
